requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.13.1
filetype==1.2.0
Flask==3.1.0
fsspec==2024.6.1
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
markdown2==2.5.3
markdownify==0.13.1
marker-pdf==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==24.2
pdftext==0.5.1
pillow==10.4.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyparsing==3.2.1
pypdfium2==4.30.0
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
soupsieve==2.6
surya-ocr==0.9.3
sympy==1.13.1
texify==0.2.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
